+++ liquid_extra/tags/__init__.py
@@ -1,1 +1,3 @@
 from .if_not import IfNotTag
+from .macro import MacroTag
+from .macro import CallTag